requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.0
sentence-transformers==3.0.1
pinecone==5.0.1
pypdf==4.3.1
groq==0.9.0
httpx==0.25.0
tqdm==4.66.5
langchain_community
streamlit==1.39.0
requests==2.32.3
pdfplumber==0.9.0
plotly
pandas